requirements.txt
langchain
python-dotenv
gpt4all
faiss-cpu
google-generativeai
langchain-google-genai